Switch to Akka HTTP based AWS connector
Removes AWS SDK
Adds Alpakka as new Client
To perform a clean change we refactored the DynamoDBJournal so that the Helper can be replaced in isolation.
Introduces settings for Backoff and Retries instead of magic numbers.
Refactored Describe from Type Class to pattern matching for better testablility and simpler code
This fixes issue #3